Add nested pgo/ library module + Fetch() convenience entry point
This is a non-breaking restructure that makes the search/download/merge
logic importable as a Go library, without changing anything about the
existing CLI.

Layout:

    .                       package main   (CLI; same behavior, same flags)
    ./pgo                   package pgo    (library, separate go.mod)

The CLI stays at the repo root, so
`go run github.com/DataDog/datadog-pgo@latest 'service:foo env:prod' default.pgo`
keeps working as before. The library lives at
github.com/DataDog/datadog-pgo/pgo with its own go.mod, so library
consumers pull only google/pprof + sourcegraph/conc, and not the CLI's
lmittmann/tint / mattn/go-isatty dependencies. The root module pulls
the library via a local `replace` directive for in-repo dev; consumers
resolve against a real tagged version once `pgo/v0.x.y` is cut.

In addition to the move, this commit adds a single high-level
convenience entry point:

    pgo.Fetch(ctx, []string{"service:foo env:prod"},
        "./cmd/foo/default.pgo",
        pgo.Options{Logger: log, SoftFail: true},
    )

Options exposes everything the CLI flags expose (APIKey/AppKey/Site to
override env-var credentials, Logger, Timeout, ProfilesPerQuery,
Window, SoftFail), so SDK callers don't have to reach for the
lower-level building blocks for the common case. Fetch falls back to
DD_API_KEY / DD_APP_KEY / DD_SITE env vars when Options fields are
empty.

Existing library symbols are exported under the same names as before:
Client, ClientFromEnv, SearchQuery/SearchFilter/SearchSort, JSONTime,
SearchProfile, MergedProfile, ProfileDownload, ProfilesDownload,
SearchDownloadMerge, ApplyNoInlineHack. Name/Version are promoted from
unexported; BuildQueries is promoted from lowercase buildQueries.

CLI behavior is otherwise unchanged: same flags (-fail, -from, -json,
-profiles, -timeout, -v), same help text, same soft-fail semantics.
`scripts/update_readme.go -mode check` still passes.

A `go.work` at the repo root lets the CI workflow build and test both
modules with `./... ./pgo/...` in a single step each (without go.work,
the toolchain skips submodules during ./... expansion).

Verified end-to-end against the live Datadog API: pgo.Fetch returned 5
profiles for a test query, merged + no-inline-hack'd them, wrote a
.pgo file, and the result round-tripped cleanly through pprof.Parse
(346 samples / 1066 locations / 407 functions across the two expected
sample types).

Environment: Datadog workspace
