+++ .github/workflows/go.yml
@@ -18,10 +18,10 @@ jobs:
         go-version: '1.21'
 
     - name: Build
-      run: go build .
+      run: go build ./... ./pgo/...
 
     - name: Test
-      run: go test -v ./...
+      run: go test -v ./... ./pgo/...
 
     - name: Check README up-to-date
       run: go run ./scripts/update_readme.go -mode check
