+++ go.work
@@ -1,0 +1,6 @@
+go 1.25.0
+
+use (
+	.
+	./pgo
+)
