+++ pgo/noinline_test.go
@@ -1,4 +1,4 @@
-package main
+package pgo
 
 import (
 	"os"
